requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
absl-py=1.0.0=pypi_0
albumentations=1.3.0=pypi_0
anyio=3.5.0=pypi_0
argon2-cffi=21.3.0=pypi_0
argon2-cffi-bindings=21.2.0=pypi_0
argparse=1.4.0=pypi_0
attrs=21.4.0=pypi_0
autocommand=2.2.2=pypi_0
babel=2.9.1=pypi_0
backcall=0.2.0=pypi_0
beautifulsoup4=4.11.1=pypi_0
blas=1.0=mkl
bleach=5.0.0=pypi_0
brotlipy=0.7.0=py37h2bbff1b_1003
ca-certificates=2022.3.29=haa95532_0
cachetools=5.0.0=pypi_0
certifi=2021.10.8=py37haa95532_2
cffi=1.15.0=py37h2bbff1b_1
chardet=3.0.4=pypi_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
cheroot=10.0.0=pypi_0
click=8.1.3=pypi_0
clip=1.0=pypi_0
cloudpickle=2.0.0=pypi_0
colorama=0.4.4=pypi_0
cryptography=36.0.0=py37h21b164f_0
cudatoolkit=11.3.1=h59b6b97_2
cycler=0.11.0=pypi_0
cython=0.29.28=py37hd77b12b_0
debugpy=1.6.0=pypi_0
decorator=5.1.1=pypi_0
defusedxml=0.7.1=pypi_0
djitellopy=2.4.0=pypi_0
docker-pycreds=0.4.0=pypi_0
entrypoints=0.4=pypi_0
et-xmlfile=1.1.0=pypi_0
fairscale=0.4.4=pypi_0
fastjsonschema=2.15.3=pypi_0
ffmpeg=1.4=pypi_0
filelock=3.12.2=pypi_0
flatbuffers=2.0=pypi_0
fonttools=4.32.0=pypi_0
freetype=2.10.4=hd328e21_0
fsspec=2023.1.0=pypi_0
ftfy=6.1.1=pypi_0
funcy=1.17=pypi_0
gitdb=4.0.9=pypi_0
gitpython=3.1.29=pypi_0
google-auth=2.6.6=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
googletrans=4.0.0rc1=pypi_0
grpcio=1.44.0=pypi_0
h11=0.9.0=pypi_0
h2=3.2.0=pypi_0
hpack=3.0.0=pypi_0
hstspreload=2023.1.1=pypi_0
httpcore=0.9.1=pypi_0
httpx=0.13.3=pypi_0
huggingface-hub=0.16.4=pypi_0
hyperframe=5.2.0=pypi_0
icc_rt=2019.0.0=h0cc432a_1
idna=2.10=pypi_0
imageio=2.22.2=pypi_0
importlib-metadata=4.11.3=pypi_0
importlib-resources=5.7.0=pypi_0
inflect=6.0.5=pypi_0
intel-openmp=2021.4.0=haa95532_3556
ipykernel=6.13.0=pypi_0
ipython=7.32.0=pypi_0
ipython-genutils=0.2.0=pypi_0
jaraco-collections=4.2.0=pypi_0
jaraco-context=4.3.0=pypi_0
jaraco-functools=3.7.0=pypi_0
jaraco-text=3.11.1=pypi_0
jedi=0.18.1=pypi_0
jinja2=3.1.1=pypi_0
joblib=1.1.0=pypi_0
jpeg=9d=h2bbff1b_0
json5=0.9.6=pypi_0
jsonschema=4.4.0=pypi_0
jupyter-client=7.2.2=pypi_0
jupyter-core=4.9.2=pypi_0
jupyter-server=1.16.0=pypi_0
jupyterlab=3.3.4=pypi_0
jupyterlab-pygments=0.2.2=pypi_0
jupyterlab-server=2.12.0=pypi_0
kiwisolver=1.4.2=pypi_0
libpng=1.6.37=h2a8f88b_0
libtiff=4.2.0=hd0e1b90_0
libuv=1.40.0=he774522_0
libwebp=1.2.2=h2bbff1b_0
lxml=5.0.0=pypi_0
lz4-c=1.9.3=h2bbff1b_1
markdown=3.3.6=pypi_0
markupsafe=2.1.1=pypi_0
matplotlib=3.5.1=pypi_0
matplotlib-inline=0.1.3=pypi_0
mistune=0.8.4=pypi_0
mkl=2021.4.0=haa95532_640
mkl-service=2.4.0=py37h2bbff1b_0
mkl_fft=1.3.1=py37h277e83a_0
mkl_random=1.2.2=py37hf11a4ad_0
more-itertools=9.1.0=pypi_0
nbclassic=0.3.7=pypi_0
nbclient=0.6.0=pypi_0
nbconvert=6.5.0=pypi_0
nbformat=5.3.0=pypi_0
nest-asyncio=1.5.5=pypi_0
networkx=2.6.3=pypi_0
nltk=3.8.1=pypi_0
notebook=6.4.10=pypi_0
notebook-shim=0.1.0=pypi_0
numpy=1.21.5=py37h7a0a035_1
numpy-base=1.21.5=py37hca35cd5_1
oauthlib=3.2.0=pypi_0
onnx=1.11.0=pypi_0
onnxruntime=1.11.0=pypi_0
opencv-python=4.8.1.78=pypi_0
opencv-python-headless=4.6.0.66=pypi_0
openpyxl=3.1.2=pypi_0
openssl=1.1.1n=h2bbff1b_0
packaging=21.3=pypi_0
pandas=1.3.5=pypi_0
pandocfilters=1.5.0=pypi_0
panopticapi=0.1=pypi_0
parso=0.8.3=pypi_0
pathtools=0.1.2=pypi_0
pdfminer-six=20231228=pypi_0
pickleshare=0.7.5=pypi_0
pillow=9.0.1=py37hdc2b20a_0
pip=21.2.4=py37haa95532_0
portend=3.1.0=pypi_0
prometheus-client=0.14.1=pypi_0
promise=2.3=pypi_0
prompt-toolkit=3.0.29=pypi_0
protobuf=3.20.0=pypi_0
psutil=5.9.0=pypi_0
py-cpuinfo=9.0.0=pypi_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pycocoevalcap=1.2=pypi_0
pycocotools=2.0.4=pypi_0
pycparser=2.21=pyhd3eb1b0_0
pydantic=1.10.13=pypi_0
pygments=2.11.2=pypi_0
pyopenssl=22.0.0=pyhd3eb1b0_0
pyparsing=3.0.8=pypi_0
pyrsistent=0.18.1=pypi_0
pysocks=1.7.1=py37_1
python=3.7.13=h6244533_0
python-dateutil=2.8.2=pypi_0
python-docx=1.1.0=pypi_0
pytorch=1.11.0=py3.7_cuda11.3_cudnn8_0
pytorch-mutex=1.0=cuda
pytz=2022.1=pypi_0
pywavelets=1.3.0=pypi_0
pywin32=303=pypi_0
pywinpty=2.0.5=pypi_0
pyyaml=6.0=pypi_0
pyzmq=22.3.0=pypi_0
qudida=0.0.4=pypi_0
regex=2023.10.3=pypi_0
requests=2.27.1=pyhd3eb1b0_0
requests-oauthlib=1.3.1=pypi_0
rfc3986=1.5.0=pypi_0
rsa=4.8=pypi_0
sacremoses=0.0.53=pypi_0
scikit-image=0.19.3=pypi_0
scikit-learn=1.0.2=pypi_0
scikit-multilearn=0.2.0=pypi_0
scipy=1.7.3=py37h0a974cb_0
seaborn=0.11.2=pypi_0
send2trash=1.8.0=pypi_0
sentry-sdk=1.10.1=pypi_0
setproctitle=1.3.2=pypi_0
setuptools=61.2.0=py37haa95532_0
sgmllib3k=1.0.0=pypi_0
shortuuid=1.0.9=pypi_0
six=1.16.0=pyhd3eb1b0_1
smmap=5.0.0=pypi_0
sniffio=1.2.0=pypi_0
soupsieve=2.3.2.post1=pypi_0
sqlite=3.38.2=h2bbff1b_0
submitit=1.4.2=pypi_0
tempora=5.3.0=pypi_0
tensorboard=2.8.0=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.1=pypi_0
terminado=0.13.3=pypi_0
thop=0.0.31-2005241907=pypi_0
threadpoolctl=3.1.0=pypi_0
tifffile=2021.11.2=pypi_0
timm=0.4.12=pypi_0
tinycss2=1.1.1=pypi_0
tk=8.6.11=h2bbff1b_0
tokenizers=0.10.3=pypi_0
torchaudio=0.11.0=py37_cu113
torchvision=0.12.0=py37_cu113
tornado=6.1=pypi_0
tqdm=4.64.0=pypi_0
traitlets=5.1.1=pypi_0
transformers=4.15.0=pypi_0
typing-extensions=4.7.1=pypi_0
ultralytics=8.0.145=pypi_0
urllib3=1.26.12=pypi_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
wandb=0.13.4=pypi_0
wcwidth=0.2.5=pypi_0
webencodings=0.5.1=pypi_0
websocket-client=1.3.2=pypi_0
werkzeug=2.1.1=pypi_0
wheel=0.37.1=pyhd3eb1b0_0
win_inet_pton=1.1.0=py37haa95532_0
wincertstore=0.2=py37haa95532_2
xz=5.2.5=h62dcd97_0
zc-lockfile=3.0.post1=pypi_0
zipp=3.8.0=pypi_0
zlib=1.2.11=hbd8134f_5
zstd=1.4.9=h19a0ad4_0
